Prototype context split
